Update time collections tests
