+++ pymeos/tests/collections/time/tstzspan_test.py
@@ -1,235 +1,272 @@
 from copy import copy
 from datetime import datetime, timezone, timedelta
-from typing import List
 
 import pytest
 
 from pymeos import (
     TsTzSpan,
     TsTzSpanSet,
+    TBox,
+    STBox,
     TFloatInst,
     TFloatSeq,
-    STBox,
     TFloatSeqSet,
-    TBox,
 )
 from tests.conftest import TestPyMEOS
 
 
-class TestPeriodSet(TestPyMEOS):
-    tstzspanset = TsTzSpanSet("{[2019-09-01, 2019-09-02], [2019-09-03, 2019-09-04]}")
+class TestTsTzSpan(TestPyMEOS):
+    tstzspan = TsTzSpan("(2019-09-08 00:00:00+0, 2019-09-10 00:00:00+0)")
 
     @staticmethod
-    def assert_tstzspanset_equality(
-        tstzspanset: TsTzSpanSet, tstzspans: List[TsTzSpan]
+    def assert_tstzspan_equality(
+        tstzspan: TsTzSpan,
+        lower: datetime = None,
+        upper: datetime = None,
+        lower_inc: bool = None,
+        upper_inc: bool = None,
     ):
-        assert tstzspanset.num_tstzspans() == len(tstzspans)
-        assert tstzspanset.tstzspans() == tstzspans
+        if lower is not None:
+            assert tstzspan.lower() == lower
+        if upper is not None:
+            assert tstzspan.upper() == upper
+        if lower_inc is not None:
+            assert tstzspan.lower_inc() == lower_inc
+        if upper_inc is not None:
+            assert tstzspan.upper_inc() == upper_inc
 
 
-class TestPeriodSetConstructors(TestPeriodSet):
-    def test_string_constructor(self):
-        self.assert_tstzspanset_equality(
-            self.tstzspanset,
-            [
-                TsTzSpan("[2019-09-01, 2019-09-02]"),
-                TsTzSpan("[2019-09-03, 2019-09-04]"),
-            ],
-        )
+class TestTsTzSpanConstructors(TestTsTzSpan):
+    @pytest.mark.parametrize(
+        "source, params",
+        [
+            (
+                "(2019-09-08 00:00:00+0, 2019-09-10 00:00:00+0)",
+                (
+                    datetime(2019, 9, 8, tzinfo=timezone.utc),
+                    datetime(2019, 9, 10, tzinfo=timezone.utc),
+                    False,
+                    False,
+                ),
+            ),
+            (
+                "[2019-09-08 00:00:00+0, 2019-09-10 00:00:00+0]",
+                (
+                    datetime(2019, 9, 8, tzinfo=timezone.utc),
+                    datetime(2019, 9, 10, tzinfo=timezone.utc),
+                    True,
+                    True,
+                ),
+            ),
+        ],
+    )
+    def test_string_constructor(self, source, params):
+        tstzspan = TsTzSpan(source)
+        self.assert_tstzspan_equality(tstzspan, *params)
 
-    def test_span_list_constructor(self):
-        tstzspanset = TsTzSpanSet(
-            span_list=[
-                TsTzSpan("[2019-09-01, 2019-09-02]"),
-                TsTzSpan("[2019-09-03, 2019-09-04]"),
-            ]
+    @pytest.mark.parametrize(
+        "input_lower,input_upper,lower,upper",
+        [
+            (
+                "2019-09-08 00:00:00+0",
+                "2019-09-10 00:00:00+0",
+                datetime(2019, 9, 8, tzinfo=timezone.utc),
+                datetime(2019, 9, 10, tzinfo=timezone.utc),
+            ),
+            (
+                datetime(2019, 9, 8, tzinfo=timezone.utc),
+                datetime(2019, 9, 10, tzinfo=timezone.utc),
+                datetime(2019, 9, 8, tzinfo=timezone.utc),
+                datetime(2019, 9, 10, tzinfo=timezone.utc),
+            ),
+            (
+                datetime(2019, 9, 8, tzinfo=timezone.utc),
+                "2019-09-10 00:00:00+0",
+                datetime(2019, 9, 8, tzinfo=timezone.utc),
+                datetime(2019, 9, 10, tzinfo=timezone.utc),
+            ),
+        ],
+        ids=["string", "datetime", "mixed"],
+    )
+    def test_constructor_bounds(self, input_lower, input_upper, lower, upper):
+        tstzspan = TsTzSpan(lower=lower, upper=upper)
+        self.assert_tstzspan_equality(tstzspan, lower, upper)
+
+    def test_constructor_bound_inclusivity_defaults(self):
+        tstzspan = TsTzSpan(lower="2019-09-08 00:00:00+0", upper="2019-09-10 00:00:00+0")
+        self.assert_tstzspan_equality(tstzspan, lower_inc=True, upper_inc=False)
+
+    @pytest.mark.parametrize(
+        "lower,upper",
+        [
+            (True, True),
+            (True, False),
+            (False, True),
+            (False, False),
+        ],
+    )
+    def test_constructor_bound_inclusivity(self, lower, upper):
+        tstzspan = TsTzSpan(
+            lower="2019-09-08 00:00:00+0",
+            upper="2019-09-10 00:00:00+0",
+            lower_inc=lower,
+            upper_inc=upper,
         )
-        self.assert_tstzspanset_equality(
-            tstzspanset,
-            [
-                TsTzSpan("[2019-09-01, 2019-09-02]"),
-                TsTzSpan("[2019-09-03, 2019-09-04]"),
-            ],
-        )
+        self.assert_tstzspan_equality(tstzspan, lower_inc=lower, upper_inc=upper)
+
+    def test_hexwkb_constructor(self):
+        assert self.tstzspan == TsTzSpan.from_hexwkb(self.tstzspan.as_hexwkb())
 
     def test_from_as_constructor(self):
-        assert self.tstzspanset == TsTzSpanSet(str(self.tstzspanset))
-        assert self.tstzspanset == TsTzSpanSet.from_wkb(self.tstzspanset.as_wkb())
-        assert self.tstzspanset == TsTzSpanSet.from_hexwkb(self.tstzspanset.as_hexwkb())
+        assert self.tstzspan == TsTzSpan(str(self.tstzspan))
+        assert self.tstzspan == TsTzSpan.from_wkb(self.tstzspan.as_wkb())
 
     def test_copy_constructor(self):
-        copied = copy(self.tstzspanset)
-        assert self.tstzspanset == copied
-        assert self.tstzspanset is not copied
+        other = copy(self.tstzspan)
+        assert self.tstzspan == other
+        assert self.tstzspan is not other
 
 
-class TestPeriodSetOutputs(TestPeriodSet):
+class TestTsTzSpanOutputs(TestTsTzSpan):
     def test_str(self):
-        assert (
-            str(self.tstzspanset)
-            == "{[2019-09-01 00:00:00+00, 2019-09-02 00:00:00+00], "
-            "[2019-09-03 00:00:00+00, 2019-09-04 00:00:00+00]}"
-        )
+        assert str(self.tstzspan) == "(2019-09-08 00:00:00+00, 2019-09-10 00:00:00+00)"
 
     def test_repr(self):
         assert (
-            repr(self.tstzspanset)
-            == "TsTzSpanSet({[2019-09-01 00:00:00+00, 2019-09-02 00:00:00+00], "
-            "[2019-09-03 00:00:00+00, 2019-09-04 00:00:00+00]})"
+            repr(self.tstzspan)
+            == "TsTzSpan((2019-09-08 00:00:00+00, 2019-09-10 00:00:00+00))"
         )
 
     def test_hexwkb(self):
-        assert self.tstzspanset == TsTzSpanSet.from_hexwkb(self.tstzspanset.as_hexwkb())
+        assert self.tstzspan == TsTzSpan.from_hexwkb(self.tstzspan.as_hexwkb())
 
 
-class TestPeriodSetConversions(TestPeriodSet):
-    def test_to_tstzspan(self):
-        assert self.tstzspanset.to_tstzspan() == TsTzSpan("[2019-09-01, 2019-09-04]")
+class TestTsTzSpanConversions(TestTsTzSpan):
+    def test_to_tstzspanset(self):
+        tstzspanset = self.tstzspan.to_spanset()
+        assert isinstance(tstzspanset, TsTzSpanSet)
+        assert tstzspanset.num_spans() == 1
+        assert tstzspanset.start_span() == self.tstzspan
 
 
-class TestPeriodSetAccessors(TestPeriodSet):
-    tstzspanset2 = TsTzSpanSet("{[2019-09-01, 2019-09-02), (2019-09-02, 2019-09-04]}")
+class TestTsTzSpanAccessors(TestTsTzSpan):
+    tstzspan2 = TsTzSpan("[2019-09-08 02:03:00+0, 2019-09-10 02:03:00+0]")
+
+    def test_lower(self):
+        assert self.tstzspan.lower() == datetime(2019, 9, 8, tzinfo=timezone.utc)
+        assert self.tstzspan2.lower() == datetime(2019, 9, 8, 2, 3, tzinfo=timezone.utc)
+
+    def test_upper(self):
+        assert self.tstzspan.upper() == datetime(2019, 9, 10, tzinfo=timezone.utc)
+        assert self.tstzspan2.upper() == datetime(2019, 9, 10, 2, 3, tzinfo=timezone.utc)
+
+    def test_lower_inc(self):
+        assert not self.tstzspan.lower_inc()
+        assert self.tstzspan2.lower_inc()
+
+    def test_upper_inc(self):
+        assert not self.tstzspan.upper_inc()
+        assert self.tstzspan2.upper_inc()
 
     def test_duration(self):
-        assert self.tstzspanset.duration() == timedelta(days=2)
-        assert self.tstzspanset.duration(True) == timedelta(days=3)
+        assert self.tstzspan.duration() == timedelta(days=2)
+        assert self.tstzspan2.duration() == timedelta(days=2)
 
-    def test_num_timestamps(self):
-        assert self.tstzspanset.num_timestamps() == 4
-        assert self.tstzspanset2.num_timestamps() == 3
-
-    def test_start_timestamp(self):
-        assert self.tstzspanset.start_timestamp() == datetime(
-            2019, 9, 1, tzinfo=timezone.utc
-        )
-
-    def test_end_timestamp(self):
-        assert self.tstzspanset.end_timestamp() == datetime(
-            2019, 9, 4, tzinfo=timezone.utc
-        )
-
-    def test_timestamp_n(self):
-        assert self.tstzspanset.timestamp_n(0) == datetime(
-            2019, 9, 1, tzinfo=timezone.utc
-        )
-        assert self.tstzspanset.timestamp_n(1) == datetime(
-            2019, 9, 2, tzinfo=timezone.utc
-        )
-        assert self.tstzspanset.timestamp_n(2) == datetime(
-            2019, 9, 3, tzinfo=timezone.utc
-        )
-        assert self.tstzspanset.timestamp_n(3) == datetime(
-            2019, 9, 4, tzinfo=timezone.utc
-        )
-
-    def test_timestamps(self):
-        assert self.tstzspanset.timestamps() == [
-            datetime(2019, 9, 1, tzinfo=timezone.utc),
-            datetime(2019, 9, 2, tzinfo=timezone.utc),
-            datetime(2019, 9, 3, tzinfo=timezone.utc),
-            datetime(2019, 9, 4, tzinfo=timezone.utc),
-        ]
-
-    def test_num_tstzspans(self):
-        assert self.tstzspanset.num_tstzspans() == 2
-
-    def test_start_tstzspan(self):
-        assert self.tstzspanset.start_tstzspan() == TsTzSpan("[2019-09-01, 2019-09-02]")
-
-    def test_end_tstzspan(self):
-        assert self.tstzspanset.end_tstzspan() == TsTzSpan("[2019-09-03, 2019-09-04]")
-
-    def test_tstzspan_n(self):
-        assert self.tstzspanset.tstzspan_n(0) == TsTzSpan("[2019-09-01, 2019-09-02]")
-        assert self.tstzspanset.tstzspan_n(1) == TsTzSpan("[2019-09-03, 2019-09-04]")
-
-    def test_tstzspans(self):
-        assert self.tstzspanset.tstzspans() == [
-            TsTzSpan("[2019-09-01, 2019-09-02]"),
-            TsTzSpan("[2019-09-03, 2019-09-04]"),
-        ]
+    def test_duration_in_seconds(self):
+        assert self.tstzspan.duration_in_seconds() == 172800
+        assert self.tstzspan2.duration_in_seconds() == 172800
 
     def test_hash(self):
-        assert hash(self.tstzspanset)
+        assert hash(self.tstzspan)
 
 
-class TestPeriodSetTransformations(TestPeriodSet):
+class TestTsTzSpanTransformations(TestTsTzSpan):
     @pytest.mark.parametrize(
         "delta,result",
         [
             (
                 timedelta(days=4),
-                [
-                    TsTzSpan("[2019-09-05 00:00:00+0, 2019-09-06 00:00:00+0]"),
-                    TsTzSpan("[2019-09-07 00:00:00+0, 2019-09-08 00:00:00+0]"),
-                ],
+                (
+                    datetime(2019, 9, 12, tzinfo=timezone.utc),
+                    datetime(2019, 9, 14, tzinfo=timezone.utc),
+                    False,
+                    False,
+                ),
             ),
             (
                 timedelta(days=-4),
-                [
-                    TsTzSpan("[2019-08-28 00:00:00+0, 2019-08-29 00:00:00+0]"),
-                    TsTzSpan("[2019-08-30 00:00:00+00, 2019-08-31 00:00:00+00]"),
-                ],
+                (
+                    datetime(2019, 9, 4, tzinfo=timezone.utc),
+                    datetime(2019, 9, 6, tzinfo=timezone.utc),
+                    False,
+                    False,
+                ),
             ),
             (
                 timedelta(hours=2),
-                [
-                    TsTzSpan("[2019-09-01 02:00:00+0, 2019-09-02 02:00:00+0]"),
-                    TsTzSpan("[2019-09-03 02:00:00+0, 2019-09-04 02:00:00+0]"),
-                ],
+                (
+                    datetime(2019, 9, 8, 2, tzinfo=timezone.utc),
+                    datetime(2019, 9, 10, 2, tzinfo=timezone.utc),
+                    False,
+                    False,
+                ),
             ),
             (
                 timedelta(hours=-2),
-                [
-                    TsTzSpan("[2019-08-31 22:00:00+0, 2019-09-01 22:00:00+0]"),
-                    TsTzSpan("[2019-09-02 22:00:00+0, 2019-09-03 22:00:00+0]"),
-                ],
+                (
+                    datetime(2019, 9, 7, 22, tzinfo=timezone.utc),
+                    datetime(2019, 9, 9, 22, tzinfo=timezone.utc),
+                    False,
+                    False,
+                ),
             ),
         ],
         ids=["positive days", "negative days", "positive hours", "negative hours"],
     )
     def test_shift(self, delta, result):
-        shifted = self.tstzspanset.shift(delta)
-        self.assert_tstzspanset_equality(shifted, result)
+        shifted = self.tstzspan.shift(delta)
+        self.assert_tstzspan_equality(shifted, *result)
 
     @pytest.mark.parametrize(
         "delta,result",
         [
             (
-                timedelta(days=3),
-                [
-                    TsTzSpan("[2019-09-01 00:00:00+0, 2019-09-02 00:00:00+0]"),
-                    TsTzSpan("[2019-09-03 00:00:00+0, 2019-09-04 00:00:00+0]"),
-                ],
+                timedelta(days=4),
+                (
+                    datetime(2019, 9, 8, tzinfo=timezone.utc),
+                    datetime(2019, 9, 12, tzinfo=timezone.utc),
+                    False,
+                    False,
+                ),
             ),
             (
-                timedelta(hours=6),
-                [
-                    TsTzSpan("[2019-09-01 00:00:00+0, 2019-09-01 02:00:00+0]"),
-                    TsTzSpan("[2019-09-01 04:00:00+0, 2019-09-01 06:00:00+0]"),
-                ],
+                timedelta(hours=2),
+                (
+                    datetime(2019, 9, 8, tzinfo=timezone.utc),
+                    datetime(2019, 9, 8, 2, tzinfo=timezone.utc),
+                    False,
+                    False,
+                ),
             ),
         ],
         ids=["days", "hours"],
     )
     def test_scale(self, delta, result):
-        scaled = self.tstzspanset.scale(delta)
-        self.assert_tstzspanset_equality(scaled, result)
+        scaled = self.tstzspan.scale(delta)
+        self.assert_tstzspan_equality(scaled, *result)
 
     def test_shift_scale(self):
-        shifted_scaled = self.tstzspanset.shift_scale(
-            timedelta(days=4), timedelta(hours=6)
-        )
-        self.assert_tstzspanset_equality(
+        shifted_scaled = self.tstzspan.shift_scale(timedelta(days=4), timedelta(hours=4))
+        self.assert_tstzspan_equality(
             shifted_scaled,
-            [
-                TsTzSpan("[2019-09-05 00:00:00+0, 2019-09-05 02:00:00+0]"),
-                TsTzSpan("[2019-09-05 04:00:00+0, 2019-09-05 06:00:00+0]"),
-            ],
+            datetime(2019, 9, 12, 0, tzinfo=timezone.utc),
+            datetime(2019, 9, 12, 4, tzinfo=timezone.utc),
+            False,
+            False,
         )
 
 
-class TestPeriodSetTopologicalPositionFunctions(TestPeriodSet):
+class TestTsTzSpanTopologicalPositionFunctions(TestTsTzSpan):
     tstzspan = TsTzSpan("(2020-01-01 00:00:00+0, 2020-01-31 00:00:00+0)")
     tstzspanset = TsTzSpanSet(
         "{(2020-01-01 00:00:00+0, 2020-01-31 00:00:00+0), (2021-01-01 00:00:00+0, 2021-01-31 00:00:00+0)}"
@@ -280,7 +317,7 @@ class TestPeriodSetTopologicalPositionFunctions(TestPeriodSet):
         ],
     )
     def test_is_adjacent(self, other):
-        self.tstzspanset.is_adjacent(other)
+        self.tstzspan.is_adjacent(other)
 
     @pytest.mark.parametrize(
         "other",
@@ -308,7 +345,7 @@ class TestPeriodSetTopologicalPositionFunctions(TestPeriodSet):
         ],
     )
     def test_is_contained_in(self, other):
-        self.tstzspanset.is_contained_in(other)
+        self.tstzspan.is_contained_in(other)
 
     @pytest.mark.parametrize(
         "other",
@@ -338,14 +375,15 @@ class TestPeriodSetTopologicalPositionFunctions(TestPeriodSet):
         ],
     )
     def test_contains(self, other):
-        self.tstzspanset.contains(other)
-        _ = other in self.tstzspanset
+        self.tstzspan.contains(other)
+        _ = other in self.tstzspan
 
     @pytest.mark.parametrize(
         "other",
         [
             tstzspan,
             tstzspanset,
+            timestamp,
             instant,
             discrete_sequence,
             stepwise_sequence,
@@ -357,6 +395,7 @@ class TestPeriodSetTopologicalPositionFunctions(TestPeriodSet):
         ids=[
             "tstzspan",
             "tstzspanset",
+            "timestamp",
             "instant",
             "discrete_sequence",
             "stepwise_sequence",
@@ -367,7 +406,7 @@ class TestPeriodSetTopologicalPositionFunctions(TestPeriodSet):
         ],
     )
     def test_overlaps(self, other):
-        self.tstzspanset.overlaps(other)
+        self.tstzspan.overlaps(other)
 
     @pytest.mark.parametrize(
         "other",
@@ -397,7 +436,7 @@ class TestPeriodSetTopologicalPositionFunctions(TestPeriodSet):
         ],
     )
     def test_is_same(self, other):
-        self.tstzspanset.is_same(other)
+        self.tstzspan.is_same(other)
 
     @pytest.mark.parametrize(
         "other",
@@ -427,7 +466,8 @@ class TestPeriodSetTopologicalPositionFunctions(TestPeriodSet):
         ],
     )
     def test_is_before(self, other):
-        self.tstzspanset.is_before(other)
+        self.tstzspan.is_before(other)
+        self.tstzspan.is_left(other)
 
     @pytest.mark.parametrize(
         "other",
@@ -457,7 +497,8 @@ class TestPeriodSetTopologicalPositionFunctions(TestPeriodSet):
         ],
     )
     def test_is_over_or_before(self, other):
-        self.tstzspanset.is_over_or_before(other)
+        self.tstzspan.is_over_or_before(other)
+        self.tstzspan.is_over_or_left(other)
 
     @pytest.mark.parametrize(
         "other",
@@ -487,7 +528,8 @@ class TestPeriodSetTopologicalPositionFunctions(TestPeriodSet):
         ],
     )
     def test_is_after(self, other):
-        self.tstzspanset.is_after(other)
+        self.tstzspan.is_after(other)
+        self.tstzspan.is_right(other)
 
     @pytest.mark.parametrize(
         "other",
@@ -517,7 +559,8 @@ class TestPeriodSetTopologicalPositionFunctions(TestPeriodSet):
         ],
     )
     def test_is_over_or_after(self, other):
-        self.tstzspanset.is_over_or_after(other)
+        self.tstzspan.is_over_or_after(other)
+        self.tstzspan.is_over_or_right(other)
 
     @pytest.mark.parametrize(
         "other",
@@ -547,10 +590,10 @@ class TestPeriodSetTopologicalPositionFunctions(TestPeriodSet):
         ],
     )
     def test_distance(self, other):
-        self.tstzspanset.distance(other)
+        self.tstzspan.distance(other)
 
 
-class TestPeriodSetSetFunctions(TestPeriodSet):
+class TestTsTzSpanSetFunctions(TestTsTzSpan):
     tstzspan = TsTzSpan("(2020-01-01 00:00:00+0, 2020-01-31 00:00:00+0)")
     tstzspanset = TsTzSpanSet(
         "{(2020-01-01 00:00:00+0, 2020-01-31 00:00:00+0), (2021-01-01 00:00:00+0, 2021-01-31 00:00:00+0)}"
@@ -563,8 +606,8 @@ class TestPeriodSetSetFunctions(TestPeriodSet):
         ids=["tstzspan", "tstzspanset", "timestamp"],
     )
     def test_intersection(self, other):
-        self.tstzspanset.intersection(other)
-        self.tstzspanset * other
+        self.tstzspan.intersection(other)
+        self.tstzspan * other
 
     @pytest.mark.parametrize(
         "other",
@@ -572,8 +615,8 @@ class TestPeriodSetSetFunctions(TestPeriodSet):
         ids=["tstzspan", "tstzspanset", "timestamp"],
     )
     def test_union(self, other):
-        self.tstzspanset.union(other)
-        self.tstzspanset + other
+        self.tstzspan.union(other)
+        self.tstzspan + other
 
     @pytest.mark.parametrize(
         "other",
@@ -581,32 +624,28 @@ class TestPeriodSetSetFunctions(TestPeriodSet):
         ids=["tstzspan", "tstzspanset", "timestamp"],
     )
     def test_minus(self, other):
-        self.tstzspanset.minus(other)
-        self.tstzspanset - other
+        self.tstzspan.minus(other)
+        self.tstzspan - other
 
 
-class TestPeriodSetComparisons(TestPeriodSet):
-    tstzspanset = TsTzSpanSet(
-        "{(2020-01-01 00:00:00+0, 2020-01-31 00:00:00+0), (2021-01-01 00:00:00+0, 2021-01-31 00:00:00+0)}"
-    )
-    other = TsTzSpanSet(
-        "{(2021-01-01 00:00:00+0, 2021-01-31 00:00:00+0), (2022-01-01 00:00:00+0, 2022-01-31 00:00:00+0)}"
-    )
+class TestTsTzSpanComparisons(TestTsTzSpan):
+    tstzspan = TsTzSpan("(2020-01-01 00:00:00+0, 2020-01-31 00:00:00+0)")
+    other = TsTzSpan("(2020-01-02 00:00:00+0, 2020-03-31 00:00:00+0)")
 
     def test_eq(self):
-        _ = self.tstzspanset == self.other
+        _ = self.tstzspan == self.other
 
     def test_ne(self):
-        _ = self.tstzspanset != self.other
+        _ = self.tstzspan != self.other
 
     def test_lt(self):
-        _ = self.tstzspanset < self.other
+        _ = self.tstzspan < self.other
 
     def test_le(self):
-        _ = self.tstzspanset <= self.other
+        _ = self.tstzspan <= self.other
 
     def test_gt(self):
-        _ = self.tstzspanset > self.other
+        _ = self.tstzspan > self.other
 
     def test_ge(self):
-        _ = self.tstzspanset >= self.other
+        _ = self.tstzspan >= self.other
